pdf-processor/requirements.txt
boto3==1.28.85
PyPDF2==3.0.1
reportlab==4.0.7
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.1
python-dotenv==1.0.0
unidecode==1.3.7
python-docx==0.8.11
